+++ src/app/account/account-analysis/select-facility-analysis-items/select-item-table/select-item-table.component.ts
@@ -8,9 +8,9 @@ import { DbChangesService } from 'src/app/indexedDB/db-changes.service';
 import { AccountdbService } from 'src/app/indexedDB/account-db.service';
 import { LoadingService } from 'src/app/core-components/loading/loading.service';
 import { AnalysisService } from 'src/app/facility/analysis/analysis.service';
-import { AnalysisValidationService } from 'src/app/facility/analysis/analysis-validation.service';
 import { SharedDataService } from 'src/app/shared/helper-services/shared-data.service';
 import { firstValueFrom } from 'rxjs';
+import { AnalysisValidationService } from 'src/app/shared/helper-services/analysis-validation.service';
 @Component({
   selector: 'app-select-item-table',
   templateUrl: './select-item-table.component.html',
