+++ src/app/facility/analysis/run-analysis/group-analysis/group-analysis-options/group-analysis-options.component.ts
@@ -7,9 +7,9 @@ import * as _ from 'lodash';
 import { AccountdbService } from 'src/app/indexedDB/account-db.service';
 import { FacilitydbService } from 'src/app/indexedDB/facility-db.service';
 import { DbChangesService } from 'src/app/indexedDB/db-changes.service';
-import { AnalysisValidationService } from '../../../analysis-validation.service';
 import { Router } from '@angular/router';
 import { UtilityMeterDatadbService } from 'src/app/indexedDB/utilityMeterData-db.service';
+import { AnalysisValidationService } from 'src/app/shared/helper-services/analysis-validation.service';
 
 @Component({
   selector: 'app-group-analysis-options',
