+++ src/app/facility/analysis/run-analysis/group-analysis/regression-model-selection/regression-model-menu/regression-model-menu.component.ts
@@ -13,8 +13,8 @@ import { CalanderizedMeter } from 'src/app/models/calanderization';
 import { AnalysisGroup, IdbAccount, IdbAnalysisItem, IdbFacility, IdbPredictorEntry, IdbUtilityMeter, IdbUtilityMeterData } from 'src/app/models/idb';
 import { RegressionModelsService } from 'src/app/shared/shared-analysis/calculations/regression-models.service';
 import * as _ from 'lodash';
-import { AnalysisValidationService } from 'src/app/facility/analysis/analysis-validation.service';
 import { SharedDataService } from 'src/app/shared/helper-services/shared-data.service';
+import { AnalysisValidationService } from 'src/app/shared/helper-services/analysis-validation.service';
 @Component({
   selector: 'app-regression-model-menu',
   templateUrl: './regression-model-menu.component.html',
