+++ src/app/models/accountAnalysis.ts
@@ -1,0 +1,9 @@
+
+export interface AccountAnalysisSetupErrors {
+    hasError: boolean,
+    missingName: boolean,
+    noGroups: boolean,
+    missingReportYear: boolean,
+    missingBaselineYear: boolean,
+    reportYearBeforeBaselineYear: boolean,
+}
