+++ src/app/shared/helper-services/analysis-validation.service.ts
